config/requirements.txt
certifi>=2025.6.15
charset-normalizer>=3.4.2
feedparser>=6.0.11
idna>=3.10
numpy>=2.3.1
python-dotenv>=1.1.1
python-telegram-bot>=22.2
pytz>=2025.2
requests>=2.32.4
sgmllib3k>=1.0.0
timezonefinder>=6.5.9
tornado>=6.5.1
typing_extensions>=4.14.1
urllib3>=2.5.0

# Essential testing (included in production for CI/CD)
pytest>=7.4.0
pytest-cov>=4.1.0

# Enhanced logging for production
python-json-logger>=2.0.7